refactor(Elements::Footing): - Elements::Footing include Helpers::LayeredReinforcedElements - creates Helpers::LayeredReinforcedElements with the aim of add methods to layered reinforced elements such as footing, slabs, or walls
