+++ lib/structuraid_core/elements/footing.rb
@@ -3,11 +3,11 @@ module StructuraidCore
     # A footing is a structural element that transfers load from a column to the soil.
     class Footing
       attr_reader :length_1, :length_2, :height, :material, :cover_lateral, :cover_top, :cover_bottom,
-                  :longitudinal_top_reinforcement_length_1,
-                  :longitudinal_bottom_reinforcement_length_1,
-                  :longitudinal_top_reinforcement_length_2,
-                  :longitudinal_bottom_reinforcement_length_2,
+                  :longitudinal_top_reinforcement_length_1, :longitudinal_bottom_reinforcement_length_1,
+                  :longitudinal_top_reinforcement_length_2, :longitudinal_bottom_reinforcement_length_2,
                   :coordinates_system
+
+      include Helpers::LayeredReinforcedElements
 
       VALID_SECTIONS = %i[length_1 length_2].freeze
 
@@ -25,17 +25,9 @@ module StructuraidCore
       #
       # @return [Footing] the footing
       def initialize(
-        length_1:,
-        length_2:,
-        height:,
-        material:,
-        cover_lateral:,
-        cover_top:,
-        cover_bottom:,
-        longitudinal_top_reinforcement_length_1: nil,
-        longitudinal_bottom_reinforcement_length_1: nil,
-        longitudinal_top_reinforcement_length_2: nil,
-        longitudinal_bottom_reinforcement_length_2: nil
+        length_1:, length_2:, height:, material:, cover_lateral:, cover_top:, cover_bottom:,
+        longitudinal_top_reinforcement_length_1: nil, longitudinal_bottom_reinforcement_length_1: nil,
+        longitudinal_top_reinforcement_length_2: nil, longitudinal_bottom_reinforcement_length_2: nil
       )
         @length_1 = length_1.to_f
         @length_2 = length_2.to_f
@@ -167,62 +159,6 @@ module StructuraidCore
 
       def aspect_ratio
         @length_1 / @length_2
-      end
-
-      def length_1_section_effective_area(above_middle:)
-        if above_middle
-          return @height - cover_bottom unless @longitudinal_top_reinforcement_length_1
-
-          @longitudinal_top_reinforcement_length_1.centroid_height
-        else
-          return @height - cover_top unless @longitudinal_bottom_reinforcement_length_1
-
-          @height - @longitudinal_bottom_reinforcement_length_1.centroid_height
-        end
-      end
-
-      def length_2_section_effective_area(above_middle:)
-        if above_middle
-          return @height - cover_bottom unless @longitudinal_top_reinforcement_length_2
-
-          @longitudinal_top_reinforcement_length_2.centroid_height
-        else
-          return @height - cover_top unless @longitudinal_bottom_reinforcement_length_2
-
-          @height - @longitudinal_bottom_reinforcement_length_2.centroid_height
-        end
-      end
-
-      def reinforcement_ratio_length_1(above_middle:)
-        if above_middle
-          return 0 if @longitudinal_top_reinforcement_length_1.empty?
-
-          reinforcement = @longitudinal_top_reinforcement_length_1.area
-        else
-          return 0 if @longitudinal_bottom_reinforcement_length_1.empty?
-
-          reinforcement = @longitudinal_bottom_reinforcement_length_1.area
-        end
-
-        section_direction = :length_1
-        effective_area = width(section_direction:) * effective_height(section_direction:, above_middle:)
-        reinforcement / effective_area
-      end
-
-      def reinforcement_ratio_length_2(above_middle:)
-        if above_middle
-          return 0 if @longitudinal_top_reinforcement_length_2.empty?
-
-          reinforcement = @longitudinal_top_reinforcement_length_2.area
-        else
-          return 0 if @longitudinal_bottom_reinforcement_length_2.empty?
-
-          reinforcement = @longitudinal_bottom_reinforcement_length_2.area
-        end
-
-        section_direction = :length_2
-        effective_area = width(section_direction:) * effective_height(section_direction:, above_middle:)
-        reinforcement / effective_area
       end
     end
   end
